Removed AT in favor of Reflection
Removed the AccessTransformer used for removing added LayerRenders and replaced with Reflection.

This is run infrequently enough, you have to load a ResourcePack or purposely hit F3+T that an AT was slightly overkill and any performance losses due to Reflection aren't noticable. Reflection has the benefit of being the way the Forge Team prefers people to handle the situation.